requirements.txt
aiohttp==3.9.1
aiofiles==23.2.1
openai>=1.35.0
python-dotenv>=1.0.1
websockets>=12.0
